move footer to shared
