+++ src/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import './App.css';
 import { Home } from './pages';
-import { Dialog, Footer } from './components/home';
-import { Nav } from './components/shared';
+import { Dialog } from './components/home';
+import { Nav, Footer } from './components/shared';
 import { Route, Routes } from "react-router-dom"
 
 function App() {
